requirements-dev.txt
-r requirements.txt
pytest-cov
versioneer==0.18
pre-commit
flake8
myst-parser;python_version>"3"  # not available on python 2, sphinx should be ran on py 3
sphinx;python_version>"3"
sphinx-rtd-theme;python_version>"3"